fix(layout): Increase top padding for sticky quiz header (#47)
On smaller screens, the content within the sticky quiz header
(#sticky-quiz-header-container) was too close to its top edge,
appearing cramped.

Also add to larger screens to prevent the quiz header from
appearing too close to the top of the page. But also used
`var(--app-padding-mobile)` because `var(--app-padding)`
is too large even for the desktop layout.